git-svn-id: file:///home/danlipsa/dan-svn-repo/src@2670 d596a3c1-de11-4375-84b4-10211c464620

+++ WidgetVtk.cpp
@@ -284,7 +284,7 @@ void WidgetVtk::CopyTransformFrom (int fromViewNumber)
 void WidgetVtk::ForAllPipelines (
     PipelineType::Enum type, boost::function <void (ViewNumber::Enum)> f)
 {
-    for (int i = 0; i < GetSettings ()->GetViewCount (); ++i)
+    for (size_t i = 0; i < GetSettings ()->GetViewCount (); ++i)
     {
 	ViewNumber::Enum viewNumber = ViewNumber::FromSizeT (i);
 	if (GetSettings ()->IsVtkView (viewNumber) &&
